backend/pyproject.toml
[project]
name = "swiftbrief-backend"
version = "0.1.0"
description = "SwiftBrief Backend - FastAPI service for future OCR capabilities"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "instructor>=1.7.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.9",
    "paddlepaddle>=3.0.0",
    "paddleocr>=3.0.0",
    "opencc-python-reimplemented>=0.1.7",
]
